feat: Add logic to package jcr pages
